Separate core coordinator functionality
