+++ coordinator_core/Cargo.toml
@@ -1,13 +1,12 @@
 [package]
-name = "coordinator-cli"
+name = "coordinator_core"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-coordinator_core = { path = "../coordinator_core" }
-
+bincode = { version = "2.0.0-rc.3", default-features = false, features = ["derive", "serde", "std"] } 
 frostsnap_core = { path = "../frostsnap_core", features = ["std"] }
 frostsnap_comms = { path = "../frostsnap_comms" }
 frostsnap_ext = { path = "../frostsnap_ext", features = ["tungstenite", "std"] }
@@ -27,4 +26,3 @@ bdk_chain = { rev = "cd096df9568a1bb7160482f7c36c807cce1b0b5d", git = "https://g
 bdk_electrum = { rev = "cd096df9568a1bb7160482f7c36c807cce1b0b5d", git = "https://github.com/evanlinjin/bdk.git"  }
 bdk_file_store = { rev = "cd096df9568a1bb7160482f7c36c807cce1b0b5d", git = "https://github.com/evanlinjin/bdk.git" }
 bdk_coin_select = { rev = "9129b43f1ee75185f3cc3a0f27edcd00df59a61d", git = "https://github.com/llfourn/bdk.git" }
-
