+++ coordinator_core/device-names.txt
