+++ coordinator_core/src/device_namer.rs
