+++ coordinator_core/src/signer.rs
